GDB-10021 - refactor tests and address review comments
