+++ test-cypress/integration/repository/ontop-repository.spec.js
@@ -1,8 +1,8 @@
 import {RepositorySteps} from "../../steps/repository-steps";
 import {OntopRepositorySteps} from "../../steps/ontop-repository-steps";
 import {ToasterSteps} from "../../steps/toaster-steps";
-import {REPOSITORIES_URL} from "../../support/repository-commands";
-import {RepositoriesStub} from "../../stubs/repositories-stub";
+import {RepositoriesStubs as RepositoryStubs} from "../../stubs/repositories/repositories-stubs";
+import {ModalDialogSteps} from "../../steps/modal-dialog-steps";
 
 describe('Ontop repositories', () => {
     let repositoryId;
@@ -140,7 +140,7 @@ describe('Ontop repositories', () => {
     });
 
     it('should create ontop repository', () => {
-        stubRepoCreationEndpoints();
+        RepositoryStubs.stubRepoCreationEndpoints(repositoryId);
         // When I visit the create ontop page
         OntopRepositorySteps.visitCreate();
         // The test connection button should be disabled
@@ -162,7 +162,7 @@ describe('Ontop repositories', () => {
 
         const hostName = 'localhost';
         // When I fill the host name
-        OntopRepositorySteps.getHostNameInput().type(hostName);
+        OntopRepositorySteps.typeHostName(hostName);
         // Then I expect url to be filed with the host name
         OntopRepositorySteps.getUrlInput().should('have.value', 'jdbc:mysql://localhost/{database}');
         // When I change the database driver for which the port field is required
@@ -170,39 +170,39 @@ describe('Ontop repositories', () => {
         // When I fill the host name
         OntopRepositorySteps.getHostNameInput().type(hostName);
         // When I fill the port field
-        OntopRepositorySteps.getPortInput().type(5423);
+        OntopRepositorySteps.typePort(5423);
         // Then I expect the URL to be filed with the host name and port
         OntopRepositorySteps.getUrlInput().should('have.value', 'jdbc:oracle:thin:@localhost:5423:{database}');
         // When I fill the database name field
-        OntopRepositorySteps.getDatabaseNameInput().type('database-name');
+        OntopRepositorySteps.typeDatabaseName('database-name');
         // Then I expect the URL to be filed with the host name and port
         OntopRepositorySteps.getUrlInput().should('have.value', 'jdbc:oracle:thin:@localhost:5423:database-name');
         // And the test connection button to be enabled
         OntopRepositorySteps.getTestConnectionButton().should('not.be.disabled');
         // And I add an OBDA file
-        RepositorySteps.clickObdaSelectFilesBtn();
-        RepositorySteps.uploadObdaFile('fixtures/ontop/university-complete.obda');
+        OntopRepositorySteps.clickObdaFileUploadButton();
+        OntopRepositorySteps.uploadObdaFile('fixtures/ontop/university-complete.obda');
         // When I click on create repository button
         OntopRepositorySteps.clickOnCreateRepositoryButton();
         // The Ontop repository should be created
         RepositorySteps.visit();
         // The repository list should contain the new repository, which can be activated
-        RepositorySteps.getRepositoryConnectionOffBtn(repositoryId).click();
+        RepositorySteps.clickRepositoryConnectionOffBtn(repositoryId);
         // When the repository is restarted
         RepositorySteps.restartRepository(repositoryId);
         RepositorySteps.confirmModal();
         // Then the correct messages are shown
         ToasterSteps.verifySuccess('Restarting repository ' + repositoryId);
         ToasterSteps.getToast().should('not.exist');
-        assertRepositoryStatus(repositoryId, "ACTIVE");
+        RepositorySteps.assertRepositoryStatus(repositoryId, "ACTIVE");
         // And the repo icon remains Ontop
         RepositorySteps.getRepoIcon('ontop').should('be.visible');
     });
 
     it('should edit ontop repository', () => {
-        stubEditOntopResponse();
-        stubSaveOntopResponse();
-        stubRepoCreationEndpoints();
+        RepositoryStubs.stubEditOntopResponse(repositoryId);
+        RepositoryStubs.stubSaveOntopResponse(repositoryId);
+        RepositoryStubs.stubRepoCreationEndpoints(repositoryId);
         // When I open the repositories view
         RepositorySteps.visit();
         // Then I should see an Ontop repository
@@ -219,7 +219,7 @@ describe('Ontop repositories', () => {
         RepositorySteps.typeURL('jdbc:oracle:thin:@localhost:5423:database-name');
         // And save the changes
         RepositorySteps.clickSaveEditedRepo();
-        RepositorySteps.clickModalOK();
+        ModalDialogSteps.clickOnConfirmButton();
         // The icon should still be Ontop
         RepositorySteps.getRepoIcon('ontop').should('be.visible');
     });
@@ -240,130 +240,4 @@ describe('Ontop repositories', () => {
         // Then I expect url to be calculated properly
         OntopRepositorySteps.getUrlInput().should('have.value', 'jdbc:snowflake://someHostName.snowflakecomputing.com:1234/?warehouse=test_database');
     });
-
-    function assertRepositoryStatus(repositoryId, status) {
-        cy.waitUntil(() =>
-            RepositorySteps.getRepositoryFromList(repositoryId)
-                .should('be.visible')
-                .find('.repository-status .text-secondary').should('contain.text', status));
-    }
-
-    const commonRepositoryParams = {
-        "propertiesFile": {
-            "name": "propertiesFile",
-            "label": "JDBC properties file",
-            "value": "repositories/ontop/ontop_jdbc.properties"
-        },
-        "lensesFile": {
-            "name": "lensesFile",
-            "label": "Lenses file",
-            "value": ""
-        },
-        "isShacl": {
-            "name": "isShacl",
-            "label": "Enable SHACL validation",
-            "value": "false"
-        },
-        "owlFile": {
-            "name": "owlFile",
-            "label": "Ontology file",
-            "value": ""
-        },
-        "member": {
-            "name": "member",
-            "label": "FedX repo members",
-            "value": []
-        },
-        "constraintFile": {
-            "name": "constraintFile",
-            "label": "Constraint file",
-            "value": ""
-        },
-        "obdaFile": {
-            "name": "obdaFile",
-            "label": "OBDA or R2RML file",
-            "value": "repositories/ontop/university-complete.obda"
-        },
-        "dbMetadataFile": {
-            "name": "dbMetadataFile",
-            "label": "DB metadata file",
-            "value": ""
-        }
-    };
-
-    function interceptRepository(id, statusCode, extraParams = {}) {
-        cy.intercept(`rest/repositories/${id}?location=`, {
-            statusCode,
-            body: {
-                "id": id,
-                "title": "",
-                "type": "ontop",
-                "sesameType": "graphdb:OntopRepository",
-                "location": "",
-                "hostName": "localhost",
-                "params": { ...commonRepositoryParams, ...extraParams }
-            }
-        });
-    }
-
-    function stubRepoCreationEndpoints() {
-        cy.intercept(REPOSITORIES_URL + 'all', {
-            statusCode: 200,
-            body: {
-                "": [
-                    {
-                        "id": repositoryId,
-                        "title": "",
-                        "uri": "http://address:5423/repositories/ontop",
-                        "externalUrl": "http://address:5423/repositories/ontop",
-                        "local": true,
-                        "type": "ontop",
-                        "sesameType": "graphdb:OntopRepository",
-                        "location": "",
-                        "readable": true,
-                        "writable": true,
-                        "unsupported": false,
-                        "state": "ACTIVE"
-                    }
-                ]
-            }
-        }).as('getMockRepositories');
-
-        cy.intercept(`rest/repositories/${repositoryId}/restart?location=*`, {
-            statusCode: 200,
-            body: {}
-        }).as('restartRepository');
-
-        cy.intercept(`/rest/autocomplete/enabled`, {
-            statusCode: 200,
-            body: {}
-        });
-
-        RepositoriesStub.stubNameSpaces(repositoryId, [{
-            "type": "literal",
-            "value": "http://jena.apache.org/ARQ/function/aggregate#"
-        }]);
-    }
-
-    function stubEditOntopResponse() {
-        interceptRepository(repositoryId, 200);
-    }
-
-    function stubSaveOntopResponse() {
-        cy.intercept(`rest/repositories/ontop/jdbc-properties?driverType=*`, {
-            statusCode: 200,
-            body: {
-                "id": repositoryId,
-                "title": "",
-                "type": "ontop",
-                "sesameType": "graphdb:OntopRepository",
-                "location": "",
-                "hostName": "localhost",
-                "driverClass": "postgresql",
-                "params": commonRepositoryParams
-            }
-        }).as('saveChanges');
-
-        cy.intercept(`/rest/repositories/${repositoryId}`, { statusCode: 200 });
-    }
 });
